Rename the controllers package to controller
