+++ internal/controller/main.go
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 
-package controllers
+package controller
 
 import (
 	"sigs.k8s.io/controller-runtime/pkg/metrics"
