+++ internal/controller/metrics.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package controllers
+package controller
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
