+++ internal/controller/suite_test.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package controllers
+package controller
 
 import (
 	"path/filepath"
